api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
